requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-multipart
pytest
pytest-asyncio
httpx
black
ruff
mypy
psycopg2-binary
python-dotenv
httpx